moved config into PrimaryProtonGunImpl
